+++ src/main/java/net/proselyte/springbootdemo/service/UserService.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+//логика работы с бд
 @Service
 public class UserService {
     private final UserRepository userRepository;
@@ -16,7 +17,7 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
-    public User findById(int id) {
+    public User findById(Long id) {
         return userRepository.findById(id).orElse(null);
     }
 
@@ -28,7 +29,7 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public void deleteById(int id) {
+    public void deleteById(Long id) {
         userRepository.deleteById(id);
     }
 }
